Abstract FakeConfig to test helpers for downstream apps
